requirements.txt
pandas
requests
sqlalchemy
psycopg2-binary
boto3
python-dotenv